MountComponentsを別のファイルに切り出した
